improved speed and readability
