+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup, find_packages
 with open("README.md", "r", encoding="utf-8") as f:
     LONG_DESCRIPTION = f.read()
 
-VERSION = '0.0.9'
+VERSION = '0.0.10'
 DESCRIPTION = 'A new transformer-based topic modeling library.'
 
 setup(
